+++ client/serial.go
@@ -126,6 +126,7 @@ func (sd *SerialDevClient) Run() error {
 			}
 		}
 
+		// make sure port is closed before we try to (re)open it
 		closePort()
 
 		if sd.config.Disable {